fwd declare TextBuffer in textStyle.h
- move text style parameters to TextStyle
